refactor(git): move GitManager into automation layer, drop lazy import
